+++ config/config_test.go
@@ -19,23 +19,41 @@ func TestReadConfig(t *testing.T) {
 
 	assert.Nil(t, err)
 
-	files := config.Filebeat.Prospectors
+	assert.Equal(t, uint64(1024), config.Filebeat.SpoolSize)
+
+	prospectors := config.Filebeat.Prospectors
 
 	// Check if multiple paths were read in
-	assert.Equal(t, 3, len(files))
+	assert.Equal(t, 3, len(prospectors))
 
 	// Check if full array can be read. Assumed that are ordered same as in config file
-	assert.Equal(t, 2, len(files[0].Paths))
-	assert.Equal(t, "/var/log/s*.log", files[0].Paths[1])
-	assert.Equal(t, "log", files[0].Input)
-	assert.Equal(t, 3, len(files[0].Fields))
-	assert.Equal(t, 1, len(files[0].Fields["review"]))
-	assert.Equal(t, "24h", files[0].IgnoreOlder)
-	assert.Equal(t, "10s", files[0].ScanFrequency)
+	assert.Equal(t, 2, len(prospectors[0].Paths))
+	assert.Equal(t, "/var/log/s*.log", prospectors[0].Paths[1])
+	assert.Equal(t, "log", prospectors[0].Input)
+	assert.Equal(t, 3, len(prospectors[0].Fields))
+	assert.Equal(t, 1, len(prospectors[0].Fields["review"]))
+	assert.Equal(t, "24h", prospectors[0].IgnoreOlder)
+	assert.Equal(t, "10s", prospectors[0].ScanFrequency)
 
-	assert.Equal(t, "stdin", files[2].Input)
-	assert.Equal(t, 0, len(files[2].Paths))
-	assert.Equal(t, "", files[1].ScanFrequency)
+	assert.Equal(t, "stdin", prospectors[2].Input)
+	assert.Equal(t, 0, len(prospectors[2].Paths))
+	assert.Equal(t, "", prospectors[1].ScanFrequency)
+}
+
+func TestReadConfig2(t *testing.T) {
+	// Tests with different params from config file
+	absPath, err := filepath.Abs("../tests/files/")
+
+	assert.NotNil(t, absPath)
+	assert.Nil(t, err)
+
+	config := &Config{}
+
+	// Reads second config file
+	err = cfgfile.Read(config, absPath+"/config2.yml")
+	assert.Nil(t, err)
+
+	assert.Equal(t, uint64(0), config.Filebeat.SpoolSize)
 }
 
 func TestGetConfigFiles_File(t *testing.T) {