+++ crawler/prospector_test.go
@@ -10,8 +10,10 @@ import (
 func TestProspectorInit(t *testing.T) {
 
 	prospectorConfig := config.ProspectorConfig{
-		ScanFrequency: "15s",
-		IgnoreOlder:   "100m",
+		ScanFrequency:       "15s",
+		IgnoreOlder:         "100m",
+		HarvesterBufferSize: 100,
+		TailOnRotate:        true,
 	}
 
 	prospector := Prospector{
@@ -25,13 +27,16 @@ func TestProspectorInit(t *testing.T) {
 	// Predefined values expected
 	assert.Equal(t, 100*time.Minute, prospector.ProspectorConfig.IgnoreOlderDuration)
 	assert.Equal(t, 15*time.Second, prospector.ProspectorConfig.ScanFrequencyDuration)
+	assert.Equal(t, 100, prospector.ProspectorConfig.HarvesterBufferSize)
+	assert.Equal(t, true, prospector.ProspectorConfig.TailOnRotate)
 }
 
 func TestProspectorInitEmpty(t *testing.T) {
 
 	prospectorConfig := config.ProspectorConfig{
-		ScanFrequency: "",
-		IgnoreOlder:   "",
+		ScanFrequency:       "",
+		IgnoreOlder:         "",
+		HarvesterBufferSize: 0,
 	}
 
 	prospector := Prospector{
@@ -58,6 +63,8 @@ func TestProspectorInitNotSet(t *testing.T) {
 	// Default values expected
 	assert.Equal(t, config.DefaultIgnoreOlderDuration, prospector.ProspectorConfig.IgnoreOlderDuration)
 	assert.Equal(t, config.DefaultScanFrequency, prospector.ProspectorConfig.ScanFrequencyDuration)
+	assert.Equal(t, config.DefaultHarvesterBufferSize, prospector.ProspectorConfig.HarvesterBufferSize)
+	assert.Equal(t, config.DefaultTailOnRotate, prospector.ProspectorConfig.TailOnRotate)
 }
 
 func TestProspectorInitScanFrequency0(t *testing.T) {