+++ harvester/harvester.go
@@ -7,13 +7,14 @@ import (
 )
 
 type Harvester struct {
-	Path        string /* the file path to harvest */
-	ProspectorConfig  config.ProspectorConfig
-	Offset      int64
-	FinishChan  chan int64
-	SpoolerChan chan *input.FileEvent
-
-	file *os.File /* the file being watched */
+	Path             string /* the file path to harvest */
+	ProspectorConfig config.ProspectorConfig
+	Offset           int64
+	FinishChan       chan int64
+	SpoolerChan      chan *input.FileEvent
+	BufferSize       int
+	TailOnRotate     bool
+	file             *os.File /* the file being watched */
 }
 
 // Interface for the different harvester types